// src/app/Shared/enums/validation-error-key-enum.ts
export enum ValidationErrorKeyEnum {
  REQUIRED = 'required',
  AR_NUM = 'AR_NUM',
  ENG_NUM = 'ENG_NUM',
  MIN_LENGTH = 'minlength',
  MAX_LENGTH = 'maxlength',
  START_AFTER_END = 'startAfterEnd',
  TIME_FROM_AFTER_TIME_TO = 'timeFromAfterTimeTo',
  EMAIL = 'EMAIL',
  STRONG_PASSWORD = 'strongPassword',
  PHONE_NUMBER = 'PHONE_NUMBER',
  NATIONAL_ID = 'NATIONAL_ID',
  POSITIVE_NUMBER = 'positiveNumber',
  INVALID_NUMBER = 'invalidNumber',
  NUMBER_RANGE = 'numberRange',
  PASSWORD = 'PASSWORD',
  SAUDI_NUMBER = 'SAUDI_NUMBER',
  EXPERATION_DATE_AFTER_START_DATE = 'beforeStart',
  MIN_NUM_1_9 = 'MIN_NUM_1_9',
  INVALID_URL = 'INVALID_URL',
}